+++ src/app/components/form/account/FormAccountSignIn.vue
@@ -1,51 +1,47 @@
 <template>
   <div>
-    <div v-if="!errorVisible" class="flex flex-col items-center gap-4">
-      <AppForm
-        :form="v$"
-        form-class="w-full"
-        :is-form-sent="isFormSent"
-        :submit-name="t('signIn')"
-        @submit.prevent="submit"
-      >
-        <FormInputEmailAddress
-          :form-input="v$.username"
-          @input="form.username = $event"
-        />
-        <!-- TODO: allow for username too -->
-        <FormInputPassword
-          :form-input="v$.password"
-          @input="form.password = $event"
-        />
-        <FormInputCaptcha
-          :form-input="v$.captcha"
-          is-centered
-          @input="form.captcha = $event"
-        />
-      </AppForm>
-      <ButtonColored
-        :aria-label="t('register')"
-        class="w-full"
-        :to="localePath('account-create')"
-        variant="secondary"
-      >
-        {{ t('register') }}
-      </ButtonColored>
-      <ButtonColored
-        :aria-label="t('passwordReset')"
-        class="w-full"
-        :to="localePath('account-password-reset-request')"
-        variant="tertiary"
-      >
-        {{ t('passwordReset') }}
-      </ButtonColored>
-    </div>
-    <ErrorView
-      v-else
-      :is-visible="errorVisible"
-      :description="errorDescription"
-      @close="onClose"
-    />
+    <AppStep v-slot="attributes" :is-active="!errorVisible">
+      <div v-bind="attributes" class="flex flex-col items-center gap-4">
+        <AppForm
+          :form="v$"
+          form-class="w-full"
+          :is-form-sent="isFormSent"
+          :submit-name="t('signIn')"
+          @submit.prevent="submit"
+        >
+          <FormInputEmailAddress
+            :form-input="v$.username"
+            @input="form.username = $event"
+          />
+          <!-- TODO: allow for username too -->
+          <FormInputPassword
+            :form-input="v$.password"
+            @input="form.password = $event"
+          />
+          <FormInputCaptcha
+            :form-input="v$.captcha"
+            is-centered
+            @input="form.captcha = $event"
+          />
+        </AppForm>
+        <ButtonColored
+          :aria-label="t('register')"
+          class="w-full"
+          :to="localePath('account-create')"
+          variant="secondary"
+        >
+          {{ t('register') }}
+        </ButtonColored>
+        <ButtonColored
+          :aria-label="t('passwordReset')"
+          class="w-full"
+          :to="localePath('account-password-reset-request')"
+          variant="tertiary"
+        >
+          {{ t('passwordReset') }}
+        </ButtonColored>
+      </div>
+    </AppStep>
   </div>
 </template>
 
@@ -55,7 +51,7 @@ import { useAuthenticateMutation } from '~~/gql/documents/mutations/account/acco
 
 const emit = defineEmits<{
   'signed-in': []
-  'error-visible': [boolean]
+  error: [boolean, string]
 }>()
 
 const fireAlert = useFireAlert()
@@ -113,7 +109,7 @@ const submit = async () => {
       errorDescription.value = result.error.message
     }
     errorVisible.value = true
-    emit('error-visible', true)
+    emit('error', true, errorDescription.value)
     return
   }
   try {
@@ -128,11 +124,6 @@ const submit = async () => {
     return
   }
   emit('signed-in')
-}
-
-const onClose = () => {
-  errorVisible.value = false
-  emit('error-visible', false)
 }
 
 // vuelidate